- Add abstract superclass CfMultisigTxOp: a CfOp that calls execTransaction() on a multisig - Migrate CfMultisendOp and CfOpSetup to use CfMultisigTxOp
